Hhung123

+++ PRL/View/QuanLiSanPhamChiTiet.cs
@@ -35,7 +35,7 @@ namespace Project_SHOE.View
             dataGridView1.Rows.Clear();
             int stt = 1;
             dataGridView1.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
-            dataGridView1.ColumnCount = 10;
+            dataGridView1.ColumnCount = 11;
             dataGridView1.Columns[0].Name = "STT";
             dataGridView1.Columns[1].Name = "Mã Sản Phẩm Chi Tiết";
             dataGridView1.Columns[2].Name = "Tên Sản Phẩm";
@@ -46,9 +46,10 @@ namespace Project_SHOE.View
             dataGridView1.Columns[7].Name = "Số Lượng";
             dataGridView1.Columns[8].Name = "Giá";
             dataGridView1.Columns[9].Name = "Nhân Viên";
+            dataGridView1.Columns[10].Name = "Trạng Thái";
             foreach (var s in data)
             {
-                dataGridView1.Rows.Add(stt++, s.IdSanphamct, s.TenSanPham, s.ChatLieu, s.Mau, s.KichThuoc, s.LoaiSanPham, s.Soluong, s.Dongia, s.IdNhanvien);
+                dataGridView1.Rows.Add(stt++, s.IdSanphamct, s.TenSanPham, s.ChatLieu, s.Mau, s.KichThuoc, s.LoaiSanPham, s.Soluong, s.Dongia, s.IdNhanvien, s.TrangThai);
             }
         }
         private void Clean()
@@ -115,6 +116,7 @@ namespace Project_SHOE.View
             cbb_SP.KeyDown += comboBox1_KeyDown;
             cbb_SP.KeyPress += comboBox1_KeyPress;
             label_NhanVien.Text = username;
+           loatData(_SanPhamctSer.Getview());
         }
         private void comboBox1_KeyPress(object sender, KeyPressEventArgs e)
         {
@@ -127,182 +129,136 @@ namespace Project_SHOE.View
         }
         private void btn_add_Click(object sender, EventArgs e)
         {
-            if (cbb_SP.Text == "")
+            // Đoạn mã hiển thị thông báo xác nhận trước khi thêm sản phẩm
+            DialogResult resultConfirm = MessageBox.Show("Bạn có muốn thêm sản phẩm này không?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (resultConfirm == DialogResult.Yes)
             {
-                MessageBox.Show("Sản Phẩm Đang Trống");
-                return;
-            }
-            else if (cbb_LoaiSP.Text == "")
-            {
-                MessageBox.Show("Loại Sản Phẩm Đang TRống");
-                return;
-            }
-            else if (cbb_ChatLieu.Text == "")
-            {
-                MessageBox.Show("Chất Liệu Đang TRống");
-                return;
-            }
-            else if (cbb_Mau.Text == "")
-            {
-                MessageBox.Show("Màu Đang TRống");
-                return;
-            }
-            else if (cbb_Size.Text == "")
-            {
-                MessageBox.Show("Size Đang TRống");
-                return;
-            }
-            else if (txt_giaban.Text == "")
-            {
-                MessageBox.Show("Giá Bán Đang Trống");
-                return;
-            }
-            else if (!int.TryParse(txt_giaban.Text, out int giaban) || giaban <= 10000 || giaban > 1000000000)
-            {
-                MessageBox.Show("Giá Bán Không Được Nhập Nhỏ Hơn 10.000 và Lớn Hơn 1.000.000.000", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                return; // Không thêm sản phẩm vào danh sách và kết thúc phương thức
-            }
-            else if (txt_soluong.Text == "")
-            {
-                MessageBox.Show("Số Lượng Đang Trống");
-                return;
-            }
-            else if (!int.TryParse(txt_soluong.Text, out int soLuong) || soLuong <= 0 || soLuong > 1000)
-            {
-                MessageBox.Show("Vui lòng nhập một số nguyên dương không vượt quá 1000 cho số lượng sản phẩm.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                return; // Không thêm sản phẩm vào danh sách và kết thúc phương thức
-            }
-            else
-            {
-                int idnhanvien = NhanVienService.GetIdNhanVien(int.Parse(label_NhanVien.Text));
-                var s1 = new Sanphamct();
-                int m = cbb_SP.SelectedIndex;
-                int n = cbb_LoaiSP.SelectedIndex;
-                int b = cbb_Mau.SelectedIndex;
-                int v = cbb_Size.SelectedIndex;
-                int c = cbb_ChatLieu.SelectedIndex;
-
-                s1.IdSanpham = _SanPhamSer.GetAllSP().ElementAt(m).IdSanpham;
-                s1.IdLoaisanpham = _SanPhamctSer.GetAllSPCT_LSP().ElementAt(n).IdLoaisanpham;
-                s1.IdMau = _SanPhamctSer.GetAllSPCT_Mau().ElementAt(b).IdMau;
-                s1.IdChatlieu = _SanPhamctSer.GetAllSPCT_ChatLieu().ElementAt(c).IdChatlieu;
-                s1.IdKichthuoc = _SanPhamctSer.GetAllSPCT_KichThuoc().ElementAt(v).IdKichthuoc;
-                s1.Soluong = Convert.ToInt32(txt_soluong.Text);
-                s1.Dongia = Convert.ToInt32(txt_giaban.Text);
-                s1.IdNhanvien = idnhanvien;
-
-                bool isDuplicate = _SanPhamctSer.CheckTrum(s1);
-                if (isDuplicate)
+                if (cbb_SP.Text == "" || cbb_LoaiSP.Text == "" || cbb_ChatLieu.Text == "" || cbb_Mau.Text == "" || cbb_Size.Text == "" || txt_giaban.Text == "" || txt_soluong.Text == "" || (!radioButton1.Checked && !radioButton2.Checked))
                 {
-                    MessageBox.Show("Sản phẩm đã tồn tại trong cơ sở dữ liệu.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    MessageBox.Show("Vui lòng điền đầy đủ thông tin sản phẩm và chọn trạng thái.");
+                    return;
+                }
+                else if (!int.TryParse(txt_giaban.Text, out int giaban) || giaban <= 10000 || giaban > 1000000000)
+                {
+                    MessageBox.Show("Giá Bán Không Được Nhập Nhỏ Hơn 10.000 và Lớn Hơn 1.000.000.000", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    return; // Không thêm sản phẩm vào danh sách và kết thúc phương thức
+                }
+                else if (!int.TryParse(txt_soluong.Text, out int soLuong) || soLuong <= 0 || soLuong > 1000)
+                {
+                    MessageBox.Show("Vui lòng nhập một số nguyên dương không vượt quá 1000 cho số lượng sản phẩm.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    return; // Không thêm sản phẩm vào danh sách và kết thúc phương thức
                 }
                 else
                 {
-                    MessageBox.Show(_SanPhamctSer.AddSPCT(s1));
-                    loatData(_SanPhamctSer.Getview());
-                    Clean();
+                    int idnhanvien = NhanVienService.GetIdNhanVien(label_NhanVien.Text);
+                    var s1 = new Sanphamct();
+                    int m = cbb_SP.SelectedIndex;
+                    int n = cbb_LoaiSP.SelectedIndex;
+                    int b = cbb_Mau.SelectedIndex;
+                    int v = cbb_Size.SelectedIndex;
+                    int c = cbb_ChatLieu.SelectedIndex;
+
+                    s1.IdSanpham = _SanPhamSer.GetAllSP().ElementAt(m).IdSanpham;
+                    s1.IdLoaisanpham = _SanPhamctSer.GetAllSPCT_LSP().ElementAt(n).IdLoaisanpham;
+                    s1.IdMau = _SanPhamctSer.GetAllSPCT_Mau().ElementAt(b).IdMau;
+                    s1.IdChatlieu = _SanPhamctSer.GetAllSPCT_ChatLieu().ElementAt(c).IdChatlieu;
+                    s1.IdKichthuoc = _SanPhamctSer.GetAllSPCT_KichThuoc().ElementAt(v).IdKichthuoc;
+                    s1.Soluong = Convert.ToInt32(txt_soluong.Text);
+                    s1.Dongia = Convert.ToInt32(txt_giaban.Text);
+                    s1.IdNhanvien = idnhanvien;
+                    s1.Trangthai = radioButton1.Checked ? "Còn Hàng" : "Hết Hàng"; // Gán trạng thái từ radio button
+
+                    bool isDuplicate = _SanPhamctSer.CheckTrum(s1);
+                    if (isDuplicate)
+                    {
+                        MessageBox.Show("Sản phẩm đã tồn tại trong cơ sở dữ liệu.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        MessageBox.Show(_SanPhamctSer.AddSPCT(s1));
+                        loatData(_SanPhamctSer.Getview());
+                        Clean();
+                    }
                 }
             }
-        }
 
+        }
 
         private void btn_update_Click(object sender, EventArgs e)
         {
-            if (cbb_SP.Text == "")
+            // Đoạn mã hiển thị thông báo xác nhận trước khi cập nhật thông tin sản phẩm chi tiết
+            DialogResult resultConfirm = MessageBox.Show("Bạn có muốn cập nhật thông tin sản phẩm chi tiết này không?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (resultConfirm == DialogResult.Yes)
             {
-                MessageBox.Show("Sản Phẩm Đang Trống");
-                return;
-            }
-            else if (cbb_LoaiSP.Text == "")
-            {
-                MessageBox.Show("Loại Sản Phẩm Đang TRống");
-                return;
-            }
-            else if (cbb_ChatLieu.Text == "")
-            {
-                MessageBox.Show("Chất Liệu Đang TRống");
-                return;
-            }
-            else if (cbb_Mau.Text == "")
-            {
-                MessageBox.Show("Màu Đang TRống");
-                return;
-            }
-            else if (cbb_Size.Text == "")
-            {
-                MessageBox.Show("Size Đang TRống");
-                return;
-            }
-            else if (txt_giaban.Text == "")
-            {
-                MessageBox.Show("Giá Bán Đang Trống");
-                return;
-            }
-            else if (!int.TryParse(txt_giaban.Text, out int giaban) || giaban <= 10000 || giaban > 1000000000)
-            {
-                MessageBox.Show("Giá Bán Không Được Nhập Nhỏ Hơn 10.000 và Lớn Hơn 1.000.000.000", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                return; // Không thêm sản phẩm vào danh sách và kết thúc phương thức
-            }
-            else if (txt_soluong.Text == "")
-            {
-                MessageBox.Show("Số Lượng Đang Trống");
-                return;
-            }
-            else if (!int.TryParse(txt_soluong.Text, out int soLuong) || soLuong <= 0 || soLuong > 1000)
-            {
-                MessageBox.Show("Vui lòng nhập một số nguyên dương không vượt quá 1000 cho số lượng sản phẩm.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                return; // Không thêm sản phẩm vào danh sách và kết thúc phương thức
-            }
-            else
-            {
-                int idnhanvien = NhanVienService.GetIdNhanVien(int.Parse(label_NhanVien.Text));
-                int n = cbb_LoaiSP.SelectedIndex;
-                int b = cbb_Mau.SelectedIndex;
-                int v = cbb_Size.SelectedIndex;
-                int c = cbb_ChatLieu.SelectedIndex;
-                int d = cbb_SP.SelectedIndex;
-                int maSP = int.Parse(txt_Idspct.Text);
-                // Khởi tạo đối tượng sản phẩm chi tiết mới
-                var updatedProduct = new Sanphamct()
+                if (cbb_SP.Text == "" || cbb_LoaiSP.Text == "" || cbb_ChatLieu.Text == "" || cbb_Mau.Text == "" || cbb_Size.Text == "" || txt_giaban.Text == "" || txt_soluong.Text == "" || (!radioButton1.Checked && !radioButton2.Checked))
                 {
-                    IdSanpham = _SanPhamSer.GetAllSP().ElementAtOrDefault(d)?.IdSanpham ?? 0,
-                    IdLoaisanpham = _SanPhamctSer.GetAllSPCT_LSP().ElementAtOrDefault(n)?.IdLoaisanpham ?? 0,
-                    IdMau = _SanPhamctSer.GetAllSPCT_Mau().ElementAtOrDefault(b)?.IdMau ?? 0,
-                    IdChatlieu = _SanPhamctSer.GetAllSPCT_ChatLieu().ElementAtOrDefault(c)?.IdChatlieu ?? 0,
-                    IdKichthuoc = _SanPhamctSer.GetAllSPCT_KichThuoc().ElementAtOrDefault(v)?.IdKichthuoc ?? 0,
-                    Soluong = Convert.ToInt32(txt_soluong.Text),
-                    Dongia = Convert.ToInt32(txt_giaban.Text),
-                    IdNhanvien = idnhanvien
-            };
-
-                bool isDuplicate = _SanPhamctSer.CheckTrumSua(maSP, updatedProduct);
-                if (isDuplicate)
+                    MessageBox.Show("Vui lòng điền đầy đủ thông tin sản phẩm.");
+                    return;
+                }
+                else if (!int.TryParse(txt_giaban.Text, out int giaban) || giaban <= 10000 || giaban > 1000000000)
                 {
-                    MessageBox.Show("Sản phẩm chi tiết đã tồn tại trong cơ sở dữ liệu.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    MessageBox.Show("Giá Bán Không Được Nhập Nhỏ Hơn 10.000 và Lớn Hơn 1.000.000.000", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    return; // Không cập nhật thông tin sản phẩm chi tiết và kết thúc phương thức
+                }
+                else if (!int.TryParse(txt_soluong.Text, out int soLuong) || soLuong <= 0 || soLuong > 1000)
+                {
+                    MessageBox.Show("Vui lòng nhập một số nguyên dương không vượt quá 1000 cho số lượng sản phẩm.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    return; // Không cập nhật thông tin sản phẩm chi tiết và kết thúc phương thức
                 }
                 else
                 {
-                    var result = _SanPhamctSer.UpdateSPCT(maSP, updatedProduct);
+                    int idnhanvien = NhanVienService.GetIdNhanVien(label_NhanVien.Text);
+                    int n = cbb_LoaiSP.SelectedIndex;
+                    int b = cbb_Mau.SelectedIndex;
+                    int v = cbb_Size.SelectedIndex;
+                    int c = cbb_ChatLieu.SelectedIndex;
+                    int d = cbb_SP.SelectedIndex;
+                    int maSP = int.Parse(txt_Idspct.Text);
 
-                    if (result == 3)
+                    // Khởi tạo đối tượng sản phẩm chi tiết mới
+                    var updatedProduct = new Sanphamct()
                     {
-                        MessageBox.Show("Cập nhật thành công");
-                        loatData(_SanPhamctSer.Getview());
-                        Clean();
+                        IdSanpham = _SanPhamSer.GetAllSP().ElementAtOrDefault(d)?.IdSanpham ?? 0,
+                        IdLoaisanpham = _SanPhamctSer.GetAllSPCT_LSP().ElementAtOrDefault(n)?.IdLoaisanpham ?? 0,
+                        IdMau = _SanPhamctSer.GetAllSPCT_Mau().ElementAtOrDefault(b)?.IdMau ?? 0,
+                        IdChatlieu = _SanPhamctSer.GetAllSPCT_ChatLieu().ElementAtOrDefault(c)?.IdChatlieu ?? 0,
+                        IdKichthuoc = _SanPhamctSer.GetAllSPCT_KichThuoc().ElementAtOrDefault(v)?.IdKichthuoc ?? 0,
+                        Soluong = Convert.ToInt32(txt_soluong.Text),
+                        Dongia = Convert.ToInt32(txt_giaban.Text),
+                        IdNhanvien = idnhanvien,
+                        Trangthai = radioButton1.Checked ? "Còn Hàng" : "Hết Hàng" // Gán trạng thái từ radio button
+                    };
+
+                    bool isDuplicate = _SanPhamctSer.CheckTrumSua(maSP, updatedProduct);
+                    if (isDuplicate)
+                    {
+                        MessageBox.Show("Sản phẩm chi tiết đã tồn tại trong cơ sở dữ liệu.", "Cảnh báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                     }
-                    else if (result == 2)
+                    else
                     {
-                        MessageBox.Show("Tên không được để trống");
-                        loatData(_SanPhamctSer.Getview());
-                        Clean();
-                    }
-                    else if (result == 1)
-                    {
-                        MessageBox.Show("Cập nhật thất bại");
-                        Clean();
+                        var result = _SanPhamctSer.UpdateSPCT(maSP, updatedProduct);
+
+                        if (result == 3)
+                        {
+                            MessageBox.Show("Cập nhật thành công");
+                            loatData(_SanPhamctSer.Getview());
+                            Clean();
+                        }
+                        else if (result == 2)
+                        {
+                            MessageBox.Show("Tên không được để trống");
+                            loatData(_SanPhamctSer.Getview());
+                            Clean();
+                        }
+                        else if (result == 1)
+                        {
+                            MessageBox.Show("Cập nhật thất bại");
+                            Clean();
+                        }
                     }
                 }
             }
+
         }
 
         private void btn_delete_Click(object sender, EventArgs e)
@@ -321,6 +277,13 @@ namespace Project_SHOE.View
                 int maSP = int.Parse(txt_Idspct.Text);
 
                 // Xóa sản phẩm
+                string errorMessage;
+                int check = _SanPhamctSer.checkspctcotronghoadon(maSP);
+                if (check == 1)
+                {
+                    MessageBox.Show("Sản phẩm đã tồn tại trong hóa đơn, không thể xóa.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
                 bool result = _SanPhamctSer.DeleteSPCT(maSP);
                 if (result)
                 {
@@ -335,8 +298,11 @@ namespace Project_SHOE.View
         }
 
 
+
         private void btn_Clean_Click(object sender, EventArgs e)
         {
+            radioButton1.Checked = true; // Chọn "Còn Hàng"
+            radioButton2.Checked = false; // Bỏ chọn "Hết Hàng"
             txt_giaban.Text = "";
             txt_soluong.Text = "";
             // Xóa lựa chọn trong ComboBox nếu có
@@ -350,6 +316,7 @@ namespace Project_SHOE.View
             btn_delete.Enabled = false;
         }
 
+
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             int index = e.RowIndex;
@@ -362,10 +329,25 @@ namespace Project_SHOE.View
             cbb_Size.Text = selectChild.Cells[5].Value.ToString();
             txt_soluong.Text = selectChild.Cells[7].Value.ToString();
             txt_giaban.Text = selectChild.Cells[8].Value.ToString();
+
+            // Kiểm tra và thiết lập trạng thái radio button
+            string trangThai = selectChild.Cells[9].Value.ToString();
+            if (trangThai == "Còn Hàng")
+            {
+                radioButton1.Checked = true; // Chọn "Còn Hàng"
+                radioButton2.Checked = false; // Bỏ chọn "Hết Hàng"
+            }
+            else
+            {
+                radioButton1.Checked = false; // Bỏ chọn "Còn Hàng"
+                radioButton2.Checked = true; // Chọn "Hết Hàng"
+            }
+
             btn_add.Enabled = false;
             btn_update.Enabled = true;
             btn_delete.Enabled = true;
         }
+
 
         private void dataGridView1_CellContentClick_1(object sender, DataGridViewCellEventArgs e)
         {
@@ -389,7 +371,7 @@ namespace Project_SHOE.View
         {
         }
         private void txt_soluong_KeyPress(object sender, KeyPressEventArgs e)
-        {
+       {
             // Kiểm tra nếu kí tự là khoảng trắng và không phải là phím Backspace
             if (e.KeyChar == ' ' && e.KeyChar != (char)Keys.Back)
             {
